requirements.txt
requests==2.28.1
beautifulsoup4==4.9.3
selenium==4.3.0
webdriver-manager==3.8.3
python-dotenv==0.21.0